// src/types.ts
export interface AirData {
  stationName: string
  pm10Grade: number | null
  pm25Grade: number | null
  khaiGrade: number | null
  so2Grade: number | null
  o3Grade: number | null
  no2Grade: number | null
  coGrade: number | null
  pm10Value: number | null
  pm25Value: number | null
  khaiValue: number | null
  so2Value: number | null
  o3Value: number | null
  no2Value: number | null
  coValue: number | null
  dataTime: string
  sidoName: string
}